Reorganized exif package
