+++ exif/tag/tag.go
@@ -31,8 +31,7 @@ type TagReader interface {
 	ReadAt(p []byte, off int64) (n int, err error)
 }
 
-//
-// NewTag creates a new Tag from
+// NewTag returns a new Tag from tagID, tagType, unitCount, valueOffset and rawValueOffset
 func NewTag(tagID ID, tagType Type, unitCount uint32, valueOffset uint32, rawValueOffset RawValueOffset) Tag {
 	return Tag{
 		TagID:          tagID,
