+++ exif/tag/tagType.go
