+++ exif/tag/tagType_test.go
