+++ exif/time.go
@@ -6,9 +6,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/evanoberholster/imagemeta/ifds"
-	"github.com/evanoberholster/imagemeta/ifds/exififd"
-	"github.com/evanoberholster/imagemeta/tag"
+	"github.com/evanoberholster/imagemeta/exif/ifds"
+	"github.com/evanoberholster/imagemeta/exif/ifds/exififd"
+	"github.com/evanoberholster/imagemeta/exif/tag"
 )
 
 // Errors for Parsing of Time
