+++ imagemeta.go
@@ -1,0 +1,2 @@
+// Package imagemeta provides functions for parsing and extracting Metadata from Images.
+package imagemeta
